+++ pytket/pytket/utils/operators.py
@@ -216,6 +216,16 @@ class QubitPauliOperator:
         for key, value in self._dict.items():
             self._dict[key] = value.subs(symbol_dict)
 
+    def get_dict(self) -> dict[QubitPauliString, Expr]:
+        """Generate a dict representation of QubitPauliOperator,
+        mapping each :py:class:`QubitPauliString` in the support
+        to its corresponding value.
+
+        :return: A dict of Pauli strings and their coefficients
+          as key-value pairs
+        """
+        return self._dict
+
     def to_list(self) -> list[dict[str, Any]]:
         """Generate a list serialized representation of QubitPauliOperator,
          suitable for writing to JSON.