+++ pytket/tests/utils_test.py
@@ -22,7 +22,7 @@ import pytest
 from hypothesis import HealthCheck, given, settings, strategies
 from hypothesis.strategies._internal import SearchStrategy
 from simulator import TketSimBackend, TketSimShotBackend  # type: ignore
-from sympy import symbols
+from sympy import I, symbols
 
 from pytket.backends.backend import Backend
 from pytket.circuit import Circuit, OpType, Qubit
@@ -88,6 +88,14 @@ def test_all_paulis() -> None:
     circs = _all_pauli_measurements(op, c)
     assert isinstance(circs, types.GeneratorType)
     assert len(list(circs)) == 3
+
+
+def test_dict_export() -> None:
+    qps1 = QubitPauliString(Qubit(0), Pauli.Y)
+    qps2 = QubitPauliString(Qubit(0), Pauli.X)
+    op = QubitPauliOperator({qps1: 1j, qps2: 0.5})
+    # Do equality check with a sympy "I" for the imaginary part
+    assert op.get_dict() == {qps1: 1.0 * I, qps2: 0.5}
 
 
 def test_shots_to_counts() -> None: