+++ pytket/docs/changelog.rst
@@ -7,6 +7,7 @@ Unreleased
 Features:
 
 * Add Python 3.13 support to the ZX module.
+* Add :py:meth:`QubitPauliOperator.get_dict` method.
 
 Fixes: